+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.3
+* Reuse controller when scaffold config update
+* Add elevation & corner radius config on SideDrawer
+* Add more example
+
 # 1.0.2
 * 3D-effect